Fix errors
- Fix playable spec information
- Fix playable class information
- Add disabel poll show (last 3)
- Fix add diferent lenguaje support
- Runable update can update one by one (by parameter)
- Fix logs permision, now set a owner
- Multiple fix in SQL file, edit to new lenguaje support.

+++ WEB-INF/src/sql/update_timeline.sql
@@ -7,6 +7,6 @@ CREATE TABLE `update_timeline` (
 
 CREATE TABLE `wow_token` (
     `last_updated_timestamp`    bigint(20) NOT NULL,
-    `price`                     int NOT NULL,
+    `price`                     bigint(20) NOT NULL,
     PRIMARY KEY(last_updated_timestamp)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
